fix: Reorder checks for N2 relation later
The N2 relation was previously checked early in the life cycle of the
charm. It meant that not having a relation to an AMF would block the
charm from doing some work that did not require that relation to be
present, like patching the Kubernetes Statefulset, configuring Multus
and providing the `fiveg_gnb_identity` data.

This last point can cause an issue where the CU starts and connects to
an unconfigured AMF, because the network could not yet be configured by
the NMS, because it did not have a GNB to configure.

This commit reorders the checks for N2 right before when it is required
to solve this problem. It can also make the deployment faster, as it can
make progress on the longer Kubernetes configuration in the background.