fix: Remove N2 from Multus management
Using N2 with macvlan when deploying on real hardware breaks with our
current setup. The recommended deployment puts the AMF LoadBalancer IP
on the management network, and that network also runs the K8s API.
Because of the way macvlan works, the container loses access to the K8s
API, breaking Pebble health checks and causing a crash loop.

This commits removes N2 from Multus management completely and uses the
default pod interface for communication over N2. This limits some
theoretical deployment scenarios, but none of those are currently being
tested. This can be reverted in the future if necessary, but will
require additional design work.